# specs/004-1-4-authenticated/contracts/api-contracts.yaml
openapi: 3.0.0
info:
  title: Authenticated Dashboard API
  version: 1.0.0
  description: API endpoints for authenticated app layout and dashboard
servers:
  - url: http://localhost:5173/api/v1
    description: Local development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
        - error
        - message
        - requestId
        - timestamp
      properties:
        error:
          type: string
          example: 'VALIDATION_ERROR'
        message:
          type: string
          example: 'Invalid request parameters'
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        details:
          type: object
...
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'UNAUTHORIZED'
            message: 'Invalid or expired token'
            requestId: '550e8400-e29b-41d4-a716-446655440000'
            timestamp: '2025-09-15T10:30:00Z'
...
paths:
...
  /projects/{projectId}/select:
    post:
      summary: Select active project
      description: Updates the active project in user session/preferences
      operationId: selectProject
      tags:
        - Projects
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
            format: uuid
          description: Project ID to select
      responses:
        '204':
          description: Project selected successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden - user doesn't have access to this project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'